+++ tests/test_bed/BedConsensus/test.py
@@ -10,9 +10,9 @@ class BedConsensus(BedConsensus):
         "binsize": 400,
         "cutoff": 0.6,
         "genome": "hg19",
-        "chrsize": Path(__file__).parent.parent.parent.joinpath(
-            "data/reference/hg19/chrom.sizes"
-        ).as_posix(),
+        "chrsize": Path(__file__)
+        .parent.parent.parent.joinpath("data/reference/hg19/chrom.sizes")
+        .as_posix(),
     }
 
 
@@ -23,32 +23,34 @@ class BedConsensus1(BedConsensus):
         "cutoff": 0.9,
         "genome": "hg19",
         "ignore_scores": [0, 1, 2],
-        "chrsize": Path(__file__).parent.parent.parent.joinpath(
-            "data/reference/hg19/chrom.sizes"
-        ).as_posix(),
+        "chrsize": Path(__file__)
+        .parent.parent.parent.joinpath("data/reference/hg19/chrom.sizes")
+        .as_posix(),
     }
 
 
 def pipeline():
-    return get_pipeline(__file__).set_starts(
-        BedConsensus,
-        BedConsensus1,
-    ).set_data(
-        [
+    return (
+        get_pipeline(__file__)
+        .set_starts(
+            BedConsensus,
+            BedConsensus1,
+        )
+        .set_data(
             [
-                Path(__file__).parent / "data" / "in1.bed",
-                Path(__file__).parent / "data" / "in2.bed",
-                Path(__file__).parent / "data" / "in3.bed",
+                [
+                    Path(__file__).parent / "data" / "in1.bed",
+                    Path(__file__).parent / "data" / "in2.bed",
+                    Path(__file__).parent / "data" / "in3.bed",
+                ]
             ]
-        ]
+        )
     )
 
 
 def testing(pipen):
     # assert pipen._succeeded
-    outfile = (
-        pipen.procs[-1].workdir.joinpath("0", "output", "in1_consensus.bed")
-    )
+    outfile = pipen.procs[-1].workdir.joinpath("0", "output", "in1_consensus.bed")
     assert outfile.is_file()